Move MySQL and DbDefaultConfig into its own namespace `Database`
