+++ _tools/cli/Command/InstallerCommand.php
@@ -11,9 +11,9 @@ use PH7\Cli\Exception\FileNotWritableException;
 use PH7\Cli\Exception\InvalidEmailException;
 use PH7\Cli\Exception\InvalidLicenseAgreementException;
 use PH7\Cli\Exception\Validation\InvalidPathException;
-use PH7\Cli\Misc\DbDefaultConfig;
+use PH7\Cli\Misc\Database\DbDefaultConfig;
+use PH7\Cli\Misc\Database\MySQL;
 use PH7\Cli\Misc\Helper;
-use PH7\Cli\Misc\MySQL;
 use PH7\Cli\Misc\SqlQuery;
 use PH7\Cli\Misc\Validation;
 use PH7\DbTableName;
